// internal/envapp/ui_src/src/ui/plugins/officialContainersRelease.generated.ts
// Code generated by scripts/generate_official_containers_plugin_release.go; DO NOT EDIT.
import type { PluginReleaseRef } from '@floegence/redevplugin-ui';

export const OFFICIAL_CONTAINERS_RELEASE_REF = {
  "source_id": "redeven-official",
  "release_metadata_ref": "plugins/com.redeven.official/com.redeven.official.containers/2.0.0/release.json",
  "release_metadata_sha256": "752b084c8bf805631fa96c3cb2c6e0372675e64bc468f11e812bdb0a6facc0a2",
  "publisher_id": "com.redeven.official",
  "plugin_id": "com.redeven.official.containers",
  "version": "2.0.0",
  "expected_hashes": {
    "package_sha256": "sha256:8ecf6c0d206ee557c5528e2192b2594b5d097912b83028d43ff1336532b06d13",
    "manifest_sha256": "sha256:f96534ca709165d0e30f6e7713a57ec0754f84f84ccadc2edc000f19dde7cc3d",
    "entries_sha256": "sha256:8a0048517719d934e52406dc6e9964d9ca165728d3e530d2c4df16f619bf17fa"
  }
} as const satisfies PluginReleaseRef;